modifies fix in r951 by using the byref param passing style
Sugested by Richard.
this fix only tested in win 32 bits; the prev with win32 and linux 64bits.

Looking with C goggles the call using byref looks better, it pass a pointer when a pointer is required, while the second form (without byref) will instantly trigger a  NOOOOO!!!!
I assume it works because ctypes does some promotion, else it should raise an exception.
From the 2.6.5 docs:   
"""
ctypes exports the byref() function which is used to pass parameters by reference. The same effect can be achieved with the pointer function, although pointer does a lot more work since it constructs a real pointer object, so it is faster to use byref() if you don’t need the pointer object in Python itself
"""

but also

"""
Usually, ctypes does strict type checking. This means, if you have POINTER(c_int) in the argtypes list of a function or as the type of a member field in a structure definition, only instances of exactly the same type are accepted. There are some exceptions to this rule, where ctypes accepts other objects. For example, you can pass compatible array instances instead of pointer types. So, for POINTER(c_int), ctypes accepts an array of c_int
"""
If ctypes strict type checking is indeed working, then an implicit, unstated conversion should take place in the call  glDeleteFramebuffersEXT(1, id) ?

Maybe the byref style is faster, maybe not.