Collision between CircleShape and AARectShape
In collision_model.py, adds the logic for detecting the collision
between CircleShape and AARectShape.

Implement also the methods distance and near_than in CShape for
different types of Cshapes.

Benchmark updated in benchmark/ to account for collision tests
between two different shapes.

Add test for AARects and Circle shapes.